`CircleCI`: cache Homebrew installation
We currently spend almost 3 minutes on every build installing `swifltint and `xcbeautify`, so this is worth optimizing.

As brought up in #2094, the cache key should have some sort of hash of what's installed.
However, we don't have a way to list the dependencies. For the time being, this should be enough. We can bump the "version" number to invalidate the cache whenever we need it to update.